--- src/refactoring/components/Admin/Product/ProductDetail.tsx
@@ -1,5 +1,0 @@
-export const ProductDetail = () => {
-    return (
-        
-    )
-}
