+++ src/refactoring/hooks/useCoupon.ts
@@ -10,3 +10,28 @@ export const useCoupons = (initialCoupons: Coupon[]) => {
 
   return { coupons, addCoupon };
 };
+
+export const useAdminCoupon = () => {
+  const [newCoupon, setNewCoupon] = useState<Coupon>({
+    name: '',
+    code: '',
+    discountType: 'percentage',
+    discountValue: 0
+  });
+
+  const adminAddCoupon = (onCouponAdd: (newCoupon: Coupon) => void) => {
+    onCouponAdd(newCoupon);
+    setNewCoupon({
+      name: '',
+      code: '',
+      discountType: 'percentage',
+      discountValue: 0
+    });
+  };
+
+  return {
+    newCoupon,
+    setNewCoupon,
+    adminAddCoupon,
+  };
+}
